Fixed matching-cat not always matching, and non-matching-cat matching after a matching-cat
Fixes bugs:
- Matching cat didn't always match.
- A non-matching cat could match, after matching with blue cat.

Cause: PlaceOnTree happened on Cat interact, while MatchMake happened on cat Tree interact.
And MatchMaker.Update() updated the catTraits.
The problem is that those three things need to happen in a specific order, but they could happen out of order.
MatchMake, which uses catTraits, can be the traits of the last cat that was on the tree, if Update() hadn't happened.
And if you stand in the right place you can get a Cat interact (which puts the cat on the tree), but no Tree interact.

Fix: Got rid of the Update() and tree interact, and just call everything, in the correct order, from PlaceOnTree().