The position at which the player holds the cat is now consistent
It was just moving the cat "up" in worldspace, so the holding position would vary. And the held cat could be so far away from the player (such as when picking up the cat from the tree) that it was outside the player's interaction sphere, so the player could no longer put the cat down. (And thus couldn't interact with anything else either.)