Fix for match success (or match fail) message getting stuck on
The timer was apparently added because without it the message would never appear because it was hidden immediately. But the timer meant that if you picked up the cat too quickly, then the message would stay on.

For now the message just displays for a set amount of time, and does not require a key-press.
Alternatively we could do the hide on cat pickup.