Player can place cat on play areas
Similar to placing cat on the cat tree.

Plan to try using placing on a play area as the means to discover info about the cat.

It's a little weird at the moment because if you put two cats on the same play area, they are in the same position, so you can't see both.