A Customer can have multiple conversations, and desires can be discovered through conversations
A conversation can reveal a desiredTrait.

The first customer now has 4 conversations. (in CustomerText.json)
The second and third conversations are set up (in the json) to reveal the customer's two desires.

For now, when you repeatedly converse with the customer, it advances through the conversations in order. And then the last one will repeat forever if you keep trying to talk to the customer. More complexity can be added later.