First step to multiple characters
I created Prefab variants for Marcus and Julia.
You can select which character to spawn by CustomerSpawner.nextCharacterSpawnIndex
This is the index within CustomerText.json. The character name in the json is used to find the corresponding prefab asset.

The prefab has a customerId which selects which character to use from CustomerText.json.
The material is set from the prefab on spawn.

Fixed the conversation to use the right customer's conversations.